chore: reorganize clients
