+++ tibiapi/endpoints/characters/service.py
@@ -2,8 +2,8 @@ import re as regex
 from typing import List
 
 import tibiapi.tools.strainer as strainer
-from tibiapi.gateway import client
 
+from .client import get_character
 from .enums import CharacterPageIdentifiers
 from .schemas import Achievements, Character, Characters
 
@@ -13,6 +13,7 @@ async def get_character(character_name: str) -> Character:
     Get a character by his name.
     Scrape the Tibia.com website to get the character's information.
     """
+
     page_tables = await _get_character_page_tables(character_name)
 
     return strainer.extract_basic_information(page_tables)
@@ -23,6 +24,7 @@ async def get_characters(character_name: str) -> List[Characters]:
     Get a list of all characters from a specific player.
     Scrape the Tibia.com website to get the character's information.
     """
+
     page_tables = await _get_character_page_tables(character_name)
 
     return strainer.extract_characters(page_tables)
@@ -33,6 +35,7 @@ async def get_achievements(character_name: str) -> List[Achievements]:
     Get achievements from a character.
     Scrape the Tibia.com website to get the character's achievements.
     """
+
     page_tables = await _get_character_page_tables(character_name)
 
     return strainer.extract_achievements(page_tables)
@@ -45,7 +48,8 @@ async def _get_character_page_tables(character_name: str):
     such as achievements, deaths, etc. All tables have the same class
     name "TableContent".
     """
-    page = await client.get_character(character_name)
+
+    page = await get_character(character_name)
 
     return page.find_all(
         "table", {"class": CharacterPageIdentifiers.TABLE_CONTENT.value})
